chore: move PopupAssetDetails components to individual  own files
